src/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
openai>=1.30.0
qdrant-client>=1.9.0
json-repair>=0.28.0
pydantic-settings>=2.3.0
google-adk>=0.3.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
